condense streaming logic into sequence svc files
